+++ src/AlyssaBuildConfig.h
@@ -31,4 +31,4 @@
 #define Compile_locales
 
 // Compile with zlib for gz encoding support.
-//#define Compile_zlib
+#define Compile_zlib